RAG・AIエージェント実践入門/9-3_handson/requirements.txt
langchain == 0.3.0
langchain-openai == 0.2.0
langgraph == 0.2.22
python-dotenv == 1.0.1